Move functions to their corresponding files
